+++ djs-bot/api/v1/src/routes/v1/routesHandler/commands.ts
@@ -1,0 +1,15 @@
+import { getBot } from '../../..';
+import { RouteHandler } from '../../../interfaces/common';
+
+const handler: RouteHandler = async (request, reply) => {
+  const bot = getBot();
+
+  return {
+    commands: bot.slash.map((command) => ({
+      name: command.name,
+      description: command.description,
+    })),
+  };
+};
+
+export default handler;
